# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the modular ServeoAPI management server.",
        "title": "ServeoAPI",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v2/docker/volumes/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a list of all volumes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "docker-volumes"
                ],
                "summary": "List Docker Volumes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/docker.VolumeInfo"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new Docker volume manually",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volumes"
                ],
                "summary": "Create a Docker Volume",
                "parameters": [
                    {
                        "description": "Volume details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docker.CreateVolumeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/docker.VolumeInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v2/docker/volumes/{name}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes a volume by name",
                "tags": [
                    "docker-volumes"
                ],
                "summary": "Delete a Docker Volume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Volume Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Force remove",
                        "name": "force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v2/files/{server}/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a file or directory",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Delete File",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server Name",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Relative path to file or directory",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v2/files/{server}/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List files and directories in a given path",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "List Files",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server Name",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Relative path",
                        "name": "path",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/files.FileInfo"
                            }
                        }
                    }
                }
            }
        },
        "/v2/files/{server}/read": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the content of a file",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Read File",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server Name",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Relative path to file",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v2/files/{server}/upload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Uploads a binary file using multipart/form-data",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Upload File",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server Name",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Relative destination directory path",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v2/files/{server}/write": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Writes raw text content to a file",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "summary": "Write File",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server Name",
                        "name": "server",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Relative path to file",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "docker.CreateVolumeRequest": {
            "type": "object",
            "properties": {
                "driver": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "docker.VolumeInfo": {
            "type": "object",
            "properties": {
                "driver": {
                    "type": "string"
                },
                "mountpoint": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "files.FileInfo": {
            "type": "object",
            "properties": {
                "is_dir": {
                    "type": "boolean"
                },
                "mod_time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "description": "relative path to the root",
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}